// src/types.ts
export interface ChatMessage {
  id: string;
  chatId: string;
  user: { nickname: string };
  text: string;
  timestamp: string;
  reactions?: { [emoji: string]: string[] }; // emoji: array of nicknames
  replyTo?: string; // messageId
}

export interface Chat {
  id: string;
  name: string;
}